+++ modules/nio/src/main/java/java/nio/FloatBuffer.java
@@ -85,7 +85,7 @@ public abstract class FloatBuffer extends Buffer implements Comparable<FloatBuff
                 if (array == null) {
                         throw new NullPointerException();
                 }
-                if (start < 0 || len < 0 || start + len > array.length) {
+                if (start < 0 || len < 0 || (long)start + (long)len > array.length) {
                         throw new IndexOutOfBoundsException();
                 }
 
@@ -294,7 +294,7 @@ public abstract class FloatBuffer extends Buffer implements Comparable<FloatBuff
 	 *                <code>remaining()</code>
 	 */
 	public FloatBuffer get(float[] dest, int off, int len) {
-        if (off < 0 || len < 0 || off + len > dest.length) {
+        if (off < 0 || len < 0 || (long)off + (long)len > dest.length) {
             throw new IndexOutOfBoundsException();
         }
         
@@ -449,7 +449,7 @@ public abstract class FloatBuffer extends Buffer implements Comparable<FloatBuff
 	 *                If no changes may be made to the contents of this buffer
 	 */
 	public FloatBuffer put(float[] src, int off, int len) {
-        if (off < 0 || len < 0 || off + len > src.length) {
+        if (off < 0 || len < 0 || (long)off + (long)len > src.length) {
             throw new IndexOutOfBoundsException();
         }
